third_party/WebKit/LayoutTests/fast/dom/htmlcollection-enumerated-properties-expected.txt

This tests verifies the enumerated properties on HTMLCollection and their order.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS htmlCollection.__proto__ is HTMLCollection.prototype
PASS htmlCollection.length is 6
PASS enumeratedProperties is expectedEnumeratedProperties
PASS successfullyParsed is true

TEST COMPLETE
